raise RuntimeError when parameter validation fails during inheritance flow
